renaming home to index.tsx removing some manual css clases
