+++ src/App.css
@@ -41,13 +41,6 @@
   height: 80vh;
 }
 
-/* Card-Section */
-.card-bg {
-  background-color: rgb(17, 180, 180);
-  border-radius: 1rem;
-  height: fit-content;
-}
-
 /* Converter-Section */
 .container-box {
   width: 100%;
